Test for listing discovered calendar names
