+++ tests/configwizard_test.py
@@ -1,7 +1,7 @@
 import click
 import pytest
 
-from khal.configwizard import validate_int
+from khal.configwizard import get_collection_names_from_vdirs, validate_int
 
 
 def test_validate_int():
@@ -10,3 +10,11 @@ def test_validate_int():
         validate_int('3', 0, 2)
     with pytest.raises(click.UsageError):
         validate_int('two', 0, 2)
+
+
+def test_default_vdir(metavdirs):
+    names = get_collection_names_from_vdirs([('found', f'{metavdirs}/**/', 'discover')])
+    assert names == [
+        'my private calendar', 'my calendar', 'public', 'home', 'public1', 'work',
+        'cfgcolor', 'cfgcolor_again', 'cfgcolor_once_more', 'dircolor', 'singlecollection',
+    ]
